refactor: Moved settings into navbar
